Cranelift codegen backend
I added cranelift backend so language now compiles to native executable
files. I changed memory sizes to reflect machine bytes. Some features
are broken/uninplemented - for example structure return throws an error.
VM uses too much memory to store data because it uses quantified variables,
so integer takes 8 cells of memory but is actually stored in only one.

+++ src/ast.rs
@@ -49,7 +49,7 @@ pub struct ExprId {
 
 #[derive(Debug, Hash, Eq, PartialEq, Clone, salsa::Update)]
 pub enum ExprData<'db> {
-    Number(i32),
+    Number(i64),
     Bool(bool),
     Binop(Binop, ExprId<'db>, ExprId<'db>),
     Unop(Unop, ExprId<'db>),
